Remove top level main.go
This was a hold-over from when the project was originally going to run
as a stand-alone service as the primary feature and offer the code API
as a secondary feature. However, the general use case is the opposite so
the main.go is removed to make room for migrating the other content into
the top level space.