Add integration tests for all runtime modes
Unfortunately, testing against the RPC server from the official lambda
SDK is not possible because it operates on the assumption that it is the
only runtime and that the "container" or VM it runs it will be closed
rather than attempting to shut down the RPC server. This resulting in a
requirement to embed a chunk of code from the lambda SDK in order to
tweak a few lines. The end result is that we can assert that a net/rpc
client of the lambda server works assuming the net/rpc protocol use does
not change in the official sdk. We'll need to check on the sdk from time
to time to ensure nothing has changed substantially.