Add ability to trigger errors in mock http mode
Add ability to document errors returned by any function.
Add special headers to trigger errors while in mock mode.

Note that only HTTP mode is supported as there is no mechanism for
conveying a trigger through vanilla lambda. If we were to build a real
or mock version of the lambda binary then the only way to interact with
it is through the real lambda API or through the real AWS API Gateway.
Neither of these enable us to add a trigger within a function or within
a function decorator because they do not convey the state that we need
without modifying the request body. As a result, this only works for
HTTP mode but can still be used to test serverfull-gateway templates.